refactor: split out the policy mappings
